GetGlobalAchiPercents & basic_response
So turns out that not all Steam Web API requests return JSON with
a .response on it, so there goes that bit of convenience.
In keeping with being "correct", I have removed the api_call's auto
-matic await(...).response and instead changed the return types of
all API calls to represent the the existence of .response using
api/basic_request.