url system in place!
Took me a while to figure out the-right-way to type these
appropriately. I'm probably going to come look back at these types
for future reference because they're pretty good! One problem is
that the language server seems to struggle resolving them because
they're so complex.

Next up is un-DRY-ing the rest of this code to make it more easily
understood. I think some types are going to be converted into
interfaces.